requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
openai
requests
python-dotenv
pydantic
